Use size_t for count

+++ tasks/baldin_a_word_count/common/include/common.hpp
@@ -8,8 +8,8 @@
 namespace baldin_a_word_count {
 
 using InType = std::string;
-using OutType = int;
-using TestType = std::tuple<std::string, int>;
+using OutType = size_t;
+using TestType = std::tuple<std::string, size_t>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace baldin_a_word_count
